feat(hkvc): implement /get_metadata handler, fix kvPair version to uint64

+++ hkvc/types.go
@@ -138,5 +138,5 @@ type directory struct {
 type kvPair struct {
 	key     string
 	value   string
-	version int
+	version uint64
 }
